ifeat(refactor): Add tests
